+++ srcs/autocomplete/autocomplite_specific_cases.c
@@ -6,20 +6,20 @@
 /*   By: bbaelor- <bbaelor-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/09 02:31:59 by bbaelor-          #+#    #+#             */
-/*   Updated: 2019/03/15 21:02:01 by bbaelor-         ###   ########.fr       */
+/*   Updated: 2019/03/20 13:12:01 by bbaelor-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "autocomplete.h"
 
-int		check_env_a_case(char c, char cb, int pos)
+int		check_env_a_case(char c, char cb, size_t pos)
 {
 	if (c == '$' || (c == '{' && pos > 0 && cb == '$'))
 		return (1);
 	return (0);
 }
 
-int		check_oth_a_case(t_line *line_info, int pos)
+int		check_oth_a_case(t_line *line_info, size_t pos)
 {
 	if (pos && line_info->buf[pos - 1] == ';')
 		return (1);
